requirements.txt
########################################################################
#  Production & Staging 依存 : アプリが直接 import するライブラリ
########################################################################
fastapi==0.110.3                # 2025-04-28 FastAPI v0.110.3 :contentReference[oaicite:0]{index=0}
uvicorn[standard]==0.34.2       # 2025-05-06 Uvicorn v0.34.2 :contentReference[oaicite:1]{index=1}
SQLAlchemy==2.0.41              # 2025-04-15 SQLAlchemy 2.0.41 :contentReference[oaicite:2]{index=2}
pydantic==2.11.5                # 2025-05-18 Pydantic 2.11.5 :contentReference[oaicite:3]{index=3}
pydantic_settings==2.9.1
python-jose[cryptography]==3.4.0 # 2025-02-18 python-jose 3.4.0 :contentReference[oaicite:4]{index=4}
python-dotenv==1.0.1            # 2025-01-29 python-dotenv 1.0.1 :contentReference[oaicite:5]{index=5}
openai==1.82.0                  # 2025-05-17 OpenAI SDK 1.82.0 
orjson==3.10.2                  # 高速 JSON シリアライザ
passlib[bcrypt]==1.7.4          # 2025-02-25 Passlib 1.7.4 :contentReference[oaicite:7]{index=7}
bcrypt==4.1.3                   # 2025-03-09 bcrypt 4.1.3 :contentReference[oaicite:8]{index=8}
httpx[http2]==0.27.0            # 2025-04-22 HTTPX 0.27.0 :contentReference[oaicite:9]{index=9}
alembic==1.13.1                 # 2025-03-19 Alembic 1.13.1 :contentReference[oaicite:10]{index=10}
asyncpg==0.29.0                 # PostgreSQL 非同期ドライバ
psycopg2-binary==2.9.9          # Alembic オフライン用 (同期)
jinja2==3.1.4                   # レポートテンプレート描画

########################################################################
#  Azure SDK                     ― データ・メッセージ・セキュリティ
########################################################################
#azure-core==2.0.2               # SDK 共通ランタイム
azure-core==1.34.0               # SDK 共通ランタイム
azure-identity==1.17.0          # マネージド ID で KeyVault 等に接続
azure-servicebus==7.10.0        # 2025-05-10 Service Bus 7.10.0 :contentReference[oaicite:11]{index=11}
azure-storage-blob==12.25.1     # 2025-04-30 Blob SDK 12.25.1 :contentReference[oaicite:12]{index=12}
azure-cosmos==4.9.0             # 2025-05-08 Cosmos DB SDK 4.9.0 :contentReference[oaicite:13]{index=13}

########################################################################
#  Observability / Security
########################################################################
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
aiofiles==23.2.1                # ファイル I/O／アップロード
python-multipart==0.0.9         # フォーム・ファイルアップロード受信
slowapi==0.1.9                  # ルート単位レートリミット (依存：limits)

########################################################################
#  Test / Dev 依存は extras に分離
########################################################################
pytest==8.2.0                   # 2025-05-02 pytest 8.2.0 :contentReference[oaicite:14]{index=14}
pytest-asyncio==0.23.6
pytest-cov==5.0.0
ruff==0.3.4                     # Linter + Formatter
pre-commit==3.7.0

sentry_sdk==2.29.1

amqp==5.3.1
billiard==4.2.1
celery==5.5.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
kombu==5.5.3
prompt-toolkit==3.0.51
python-dateutil==2.9.0.post0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
